requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
kuzu==0.11.3
letta
chromadb>=0.4.0
httpx>=0.24.0
pydantic>=2.0
aiosqlite>=0.19.0
tiktoken>=0.5.0
pyyaml>=6.0
sentence-transformers
